Move CallResponseTest to shared
